Move function into separate file
